rename aux statements to auxiliary
