+++ src/bloqade/stim/__init__.py
@@ -1,6 +1,6 @@
 from .groups import main as main
 from ._wrappers import *  # noqa: F403
-from .dialects.aux import *  # noqa F403
 from .dialects.gate import *  # noqa F403
 from .dialects.noise import *  # noqa F403
 from .dialects.collapse import *  # noqa F403
+from .dialects.auxiliary import *  # noqa F403
