+++ src/bloqade/stim/dialects/__init__.py
@@ -1,5 +1,5 @@
-from . import aux as aux, gate as gate, noise as noise, collapse as collapse
-from .aux.stmts import *  # noqa F403
+from . import gate as gate, noise as noise, collapse as collapse, auxiliary as auxiliary
 from .gate.stmts import *  # noqa F403
 from .noise.stmts import *  # noqa F403
 from .collapse.stmts import *  # noqa F403
+from .auxiliary.stmts import *  # noqa F403
