+++ src/bloqade/stim/dialects/auxiliary/__init__.py
