+++ src/bloqade/stim/dialects/auxiliary/_dialect.py
