+++ src/bloqade/stim/dialects/auxiliary/emit.py
