+++ src/bloqade/stim/dialects/auxiliary/lowering.py
